+++ internal/provider/data_source_catalystcenter_pnp_device.go
@@ -23,14 +23,10 @@ import (
 	"fmt"
 	"net/url"
 
-	"github.com/hashicorp/terraform-plugin-framework-validators/datasourcevalidator"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/path"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	cc "github.com/netascode/go-catalystcenter"
-	"github.com/tidwall/gjson"
 )
 
 // End of section. //template:end imports
@@ -63,13 +59,11 @@ func (d *PnPDeviceDataSource) Schema(ctx context.Context, req datasource.SchemaR
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
 				MarkdownDescription: "The id of the object",
-				Optional:            true,
 				Computed:            true,
 			},
 			"serial_number": schema.StringAttribute{
 				MarkdownDescription: "Device serial number",
-				Optional:            true,
-				Computed:            true,
+				Required:            true,
 			},
 			"stack": schema.BoolAttribute{
 				MarkdownDescription: "Device is a stacked switch",
@@ -84,14 +78,6 @@ func (d *PnPDeviceDataSource) Schema(ctx context.Context, req datasource.SchemaR
 				Computed:            true,
 			},
 		},
-	}
-}
-func (d *PnPDeviceDataSource) ConfigValidators(ctx context.Context) []datasource.ConfigValidator {
-	return []datasource.ConfigValidator{
-		datasourcevalidator.ExactlyOneOf(
-			path.MatchRoot("id"),
-			path.MatchRoot("serial_number"),
-		),
 	}
 }
 
@@ -117,31 +103,9 @@ func (d *PnPDeviceDataSource) Read(ctx context.Context, req datasource.ReadReque
 	}
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Read", config.Id.String()))
-	if config.Id.IsNull() && !config.SerialNumber.IsNull() {
-		res, err := d.client.Get(config.getPath())
-		if err != nil {
-			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve objects, got error: %s", err))
-			return
-		}
-		if value := res; len(value.Array()) > 0 {
-			value.ForEach(func(k, v gjson.Result) bool {
-				if config.SerialNumber.ValueString() == v.Get("deviceInfo.serialNumber").String() {
-					config.Id = types.StringValue(v.Get("id").String())
-					tflog.Debug(ctx, fmt.Sprintf("%s: Found object with serialNumber '%v', id: %v", config.Id.String(), config.SerialNumber.ValueString(), config.Id.String()))
-					return false
-				}
-				return true
-			})
-		}
-
-		if config.Id.IsNull() {
-			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to find object with serialNumber: %s", config.SerialNumber.ValueString()))
-			return
-		}
-	}
 
 	params := ""
-	params += "/" + url.QueryEscape(config.Id.ValueString())
+	params += "?serialNumber=" + url.QueryEscape(config.SerialNumber.ValueString())
 	res, err := d.client.Get(config.getPath() + params)
 	if err != nil {
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object, got error: %s", err))